Introduce TypeUtils to handle Type checks and operations
